# Annotations: Optimize Get queries