@pstibrany 2025-08-07
> Hi @pstibrany, whenever you have a chance, could you please take a look at this PR? Thanks!

I've reassigned the bug to the correct squad.